oss-o11y/python/frontend/requirements.txt
appdirs==1.4.4
argcomplete==1.12.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==7.1.2
colorlog==4.7.2
distlib==0.3.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-trace==0.24.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
nox==2020.12.31
numpy==1.20.2
opentelemetry-api==1.10a0
opentelemetry-exporter-gcp-trace==1.0.0
opentelemetry-exporter-prometheus==1.10a0
opentelemetry-sdk==1.10a0
packaging==21.3
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.7
protobuf==3.19.2
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.15.0
urllib3==1.26.8
virtualenv==20.4.2
Werkzeug==1.0.1
